Updated code for discord.js 12.2.0 and discord-akairo 8.0.0
Some other small changes may have been made, more details in https://github.com/JosephLGibson/Pizza-Bot/commit/d9714a3651dc03dd341c3a3ac65b24d33da714e7 and a few commits after.

+++ commands/ask.js
@@ -1,26 +1,36 @@
 const { Command } = require("discord-akairo");
 const config = require("../config.js");
 
+const commandInfo = {
+	id: "ask",
+	aliases: ["yesno"],
+	args: [{id: "message", type: "string", default: "", match: "content"}],
+	description: {
+		short: "Ask a yes/no question.",
+		extend: "",
+	}
+}
+
+commandInfo.aliases.unshift(commandInfo.id)
+commandInfo.description.long = commandInfo.description.short + "\n" + commandInfo.description.extend
+commandInfo.description.args = commandInfo.args.map(item => item.id)
+
 class AskCommand extends Command {
 	constructor() {
 		super(
-			"ask", 
-			{
-				aliases: ["ask", "yesno"],
-				args: [{id: "message", type: "string", default: "", match: "content"}],
-				description: "Ask a yes/no question."
-			}
-		)
+			commandInfo.id,
+			commandInfo
+		);
 	}
+
 	async exec(message, args) {
-		let end;
 		if (args.message == ""){
 			let rep = await message.reply("can't ask an empty question!")
-			await rep.delete(5000);
+			await rep.delete({timeout: 5000});
 		} else {
 			let name = message.member.user.username + "#" + message.member.user.discriminator
 			let start;
-			if (message.member.nickname == null ) {
+			if (message.member.nickname == null) {
 				start = name
 			} else {
 				start = message.member.nickname + ` (${name})`
@@ -29,7 +39,9 @@ class AskCommand extends Command {
 			await sent.react(config.thumbs_up);
 			await sent.react(config.thumbs_down);
 		}
-		await message.delete();
+		if (message.channel.type != "dm") {
+			return await message.delete();
+		}
 	}
 }
